Refactor: 동네인증 네비게이션
